+++ src/pages/member/GetMemberProfile.jsx
@@ -1,148 +1,63 @@
-import React, { useEffect, useState } from "react";
-import {
-  Container,
-  Row,
-  Col,
-  Card,
-  Spinner,
-  Alert,
-  Button,
-} from "react-bootstrap";
+import React, { useState } from "react";
+import { Container, Row, Col, Button } from "react-bootstrap";
+import "bootstrap/dist/css/bootstrap.min.css";
+import { useSearchParams } from "react-router-dom";
+import UseProfileFetchFeeds from "./component/profile/UseProfileFetchFeeds";
+import UseProfileInfiniteScroll from "./component/profile/UseProfileInfiniteScroll";
+import ProfileFeedList from "./component/profile/ProfileFeedList";
+import ProfileHeader from "./component/profile/ProfileHeader";
+import ProfileBody from "./component/profile/ProfileBody";
+import UseProfileMemberRelationshipList from "./component/profile/UseProfileMemberRelationshipList";
+import UseProfileBuildingSubscriptions from "./component/profile/UseProfileBuildingSubscriptions";
+import UseProfile from "./component/profile/UseProfile";
 import LogoutForm from "./component/LogoutForm";
-import { FaFacebook, FaInstagram, FaTwitter } from "react-icons/fa";
-import { getMemberProfile } from "./function/memberAxios";
-import { useSelector } from "react-redux";
-import { decryptWithLv, decryptWithLvWithUri } from "../../util/crypto";
-import { useParams } from "react-router-dom";
-
 const GetMemberProfile = () => {
-  console.log("#### GetMemberProfile 렌더링");
-
-  const authorization = useSelector((state) => state.auth.authorization);
-  const fromId = useSelector((state) => state.auth.member.memberId);
-  const { secretId, secretIv } = useParams();
-
-  const [toId, setToId] = useState(null);
-  const [profile, setProfile] = useState(null);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState(null);
-
-  useEffect(() => {
-    // console.log("URL Parameters - encryptedToId:", secretId, "IV:", secretIv);
-    if (secretId && secretIv) {
-      const decryptedToId = decryptWithLvWithUri(secretId, secretIv);
-    //   console.log("Decrypted toId: ", decryptedToId);
-      setToId(decryptedToId);
-    }
-  }, [secretId, secretIv]);
-
-  useEffect(() => {
-    if (authorization && fromId && toId) {
-      console.log("Fetching member profile data","Authorization:",authorization,"From ID:",fromId,"To ID:",toId);
-      
-      const fetchMemberProfile = async () => {
-        setLoading(true);
-        setError(null);
-        try {
-          const response = await getMemberProfile(fromId, toId);
-          console.log("Profile data:", response);
-          setProfile(response);
-        } catch (error) {
-          console.error("Error fetching member profile:", error);
-          setError("Error fetching member profile: " + error.message);
-        } finally {
-          setLoading(false);
-        }
-      };
-      fetchMemberProfile();
-    }
-  }, [authorization, toId, fromId]);
+  const { profile, toId, initialPage, isDenied } = UseProfile();
+  const { feeds, hasMore, setPage } = UseProfileFetchFeeds(toId, initialPage);
+  const { buildingSubscriptionCount } = UseProfileBuildingSubscriptions({
+    toId,
+  });
+  const { followerCount, followingCount } = UseProfileMemberRelationshipList();
+  const lastFeedElementRef = UseProfileInfiniteScroll(hasMore, setPage);
 
   return (
-    <Container className="mt-5">
-      <LogoutForm />
-      <h1 className="text-center">Member Profile</h1>
-      {loading ? (
-        <Row className="justify-content-center">
-          <Spinner animation="border" />
-        </Row>
-      ) : error ? (
-        <Alert variant="danger">{error}</Alert>
-      ) : profile ? (
-        <Card>
-          <Card.Body>
-            <Card.Title>{profile.nickname}</Card.Title>
-            <Card.Subtitle className="mb-2 text-muted">
-              Member ID: {profile.memberId}
-            </Card.Subtitle>
-            <Card.Text>
-              <p>DajungScore: {profile.dajungScore}</p>
-              <p>Profile intro: {profile.profileIntro}</p>
-              {profile.profilePhotoUrl && (
-                <img
-                  src={profile.profilePhotoUrl}
-                  alt="Profile"
-                  className="img-fluid mb-2"
-                />
-              )}
-            </Card.Text>
+    <Container
+      fluid
+      className="vh-100 w-100 d-flex flex-column justify-content-start align-items-center"
+      style={{ paddingTop: "50px", maxWidth: "100%" }}
+    >
+      <Row className="justify-content-center w-100">
+        <Col xs={12} sm={8} md={6} lg={8}>
+          <ProfileHeader />
+          {isDenied ? (
             <div>
-              <h2>feedDtoList:</h2>
-              {profile.feedDtoList &&
-                profile.feedDtoList.map((feed) => (
-                  <Card key={feed.feedId} className="mb-3">
-                    <Card.Body>
-                      <Card.Title>{feed.title}</Card.Title>
-                      <Card.Subtitle className="mb-2 text-muted">
-                        빌딩이름: {feed.buildingName}
-                      </Card.Subtitle>
-                      <Card.Text>
-                        <p>피드내용: {feed.feedText}</p>
-                        <p>작성자: {feed.writerNickname}</p>
-                        <p>작성 시간: {feed.writtenTime}</p>
-                        {feed.feedAttachementURL && (
-                          <img
-                            src={feed.feedAttachementURL}
-                            alt="첨부 이미지"
-                            className="img-fluid"
-                          />
-                        )}
-                      </Card.Text>
-                    </Card.Body>
-                  </Card>
-                ))}
+                isDenied가 true입니다.
+              <ProfileBody />
+              <Button variant="primary" className="mt-3">
+                미정
+              </Button>
+              <ProfileFeedList lastFeedElementRef={lastFeedElementRef} />
             </div>
-          </Card.Body>
-        </Card>
-      ) : (
-        <Alert variant="warning">No profile data available.</Alert>
-      )}
-      <div className="text-center mt-4">
-        <Button
-          variant="link"
-          href="https://www.facebook.com"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <FaFacebook size={30} />
-        </Button>
-        <Button
-          variant="link"
-          href="https://www.twitter.com"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <FaTwitter size={30} />
-        </Button>
-        <Button
-          variant="link"
-          href="https://www.instagram.com"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <FaInstagram size={30} />
-        </Button>
-      </div>
+          ) : (
+            <div>
+                isDenied가 false입니다.
+              <ProfileBody
+                profile={profile}
+                feeds={feeds}
+                buildingSubscriptionCount={buildingSubscriptionCount}
+                followerCount={followerCount}
+                followingCount={followingCount}
+              />
+                <LogoutForm />
+              <ProfileFeedList
+                feeds={feeds}
+                lastFeedElementRef={lastFeedElementRef}
+              />
+            </div>
+          )}
+        </Col>
+      </Row>
+      {/* Footer Component */}
     </Container>
   );
 };
