+++ src/routes/AppRoutes.jsx
@@ -14,7 +14,6 @@ import AddMemberResult from "../pages/member/AddMemberResult";
 import UpdatePwd from "../pages/member/UpdatePwd";
 import UpdatePwdResult from "../pages/member/UpdatePwdResult";
 import GetMember from "../pages/member/GetMember";
-import GetMemberProfile from "../pages/member/GetMemberProfile";
 import UpdateMember from "../pages/member/UpdateMember";
 import GetMemberRelationshipList from "../pages/member/GetMemberRelationshipList";
 
@@ -56,13 +55,14 @@ import FeedDetail from "../pages/feed/FeedDetailPage";
 import FeedForm from "../pages/feed/FeedFormPage";
 import GetMemberId from "../pages/member/GetMemberId";
 import IdFormToUpdatePwd from "../pages/member/IdFormToUpdatePwd";
-import TestComponent from '../pages/member/component/test/TestComponent';
-import Postcode from '../pages/member/component/Postcode';
-import { CSSTransition, TransitionGroup } from 'react-transition-group';
-import '../assets/css/animation.css'
-import LoginForm from '../pages/member/component/LoginForm';
-import FeedChartPage from '../pages/feed/FeedChartPage';
-import FeedBuildingListPage from '../pages/feed/FeedBuildingListPage';
+import TestComponent from "../pages/member/component/test/TestComponent";
+import Postcode from "../pages/member/component/Postcode";
+import { CSSTransition, TransitionGroup } from "react-transition-group";
+import "../assets/css/animation.css";
+import LoginForm from "../pages/member/component/LoginForm";
+import FeedChartPage from "../pages/feed/FeedChartPage";
+import FeedBuildingListPage from "../pages/feed/FeedBuildingListPage";
+import GetMemberProfile from "../pages/member/GetMemberProfile";
 const AppRoutes = () => {
   const location = useLocation();
   const navigationType = useNavigationType();
@@ -93,14 +93,6 @@ const AppRoutes = () => {
       >
         <Routes location={location}>
           <Route path="/b" element={<B />}></Route>
-          <Route
-            path="/GetMemberProfile"
-            element={
-              <GuestRoute>
-                <GetMemberProfile />
-              </GuestRoute>
-            }
-          ></Route>
           <Route path="/testComponent" element={<TestComponent />}></Route>
           <Route path="/member">
             <Route
@@ -236,11 +228,14 @@ const AppRoutes = () => {
           <Route path="/getBuildingProfile" element={<Building />} />
           <Route path="/customerSupport">
             <Route path="" element={<GetCustomerSupport />} />
-            <Route path="getChatbot" element={
+            <Route
+              path="getChatbot"
+              element={
                 <PrivateRoute>
-                    <GetChatbot />
+                  <GetChatbot />
                 </PrivateRoute>
-                } />
+              }
+            />
             <Route path="getNoticeList" element={<ListNotice />} />
             <Route path="getNotice" element={<GetNotice />} />
             <Route path="addNoticeView" element={<AddNoticeView />} />
@@ -250,14 +245,14 @@ const AppRoutes = () => {
             <Route path="listImages" element={<ListImages />} />
             <Route path="getImage" element={<GetImage />} />
             <Route path="deleteBadFeed" element={<DeleteBadFeed />} />
-        </Route>
-        <Route path="/setting" element={<MemberSetting />} />
-        <Route path="/feed">
+          </Route>
+          <Route path="/setting" element={<MemberSetting />} />
+          <Route path="/feed">
             <Route path="list" element={<FeedList />} />
-            <Route path="list/building" element={<FeedBuildingListPage/>}/>
+            <Route path="list/building" element={<FeedBuildingListPage />} />
             <Route path="detail" element={<FeedDetail />} />
             <Route path="form" element={<FeedForm />} />
-            <Route path="chart" element={<FeedChartPage/>}/>
+            <Route path="chart" element={<FeedChartPage />} />
             <Route path="" element={<FeedPages />} />
           </Route>
         </Routes>
